# feat: wasm simd support