--- a/package.json
+++ b/package.json
@@ -43,7 +43,7 @@
     "llparse-test-fixture": "^5.0.1",
     "mocha": "^9.2.2",
     "ts-node": "^9.0.0",
-    "typescript": "^4.0.3"
+    "typescript": "^5.0.3"
   },
   "dependencies": {
     "debug": "^4.2.0",