This makes SIMD version of llhttp 12% faster in [undici](https://github.com/nodejs/undici).

Example of generated code: https://gist.github.com/indutny/60dcb70611aa2b5c22ad14604a008d82

Benchmark tool: https://gist.github.com/indutny/5edb0999bfbbe6d9e7cccba35ae17950

Before:
```
generic 1150616 req/sec
simd 1134513 req/sec
```

After:
```
generic 1116794 req/sec
simd 1256336 req/sec
```

cc @mcollina @nodejs/llhttp